Add files via upload
Experimentation with Binary Logistic Regression: (NOTE: removed dropout layer for this training session)

Added the Sigmoid class as an activation function to the layers_activators module and applied it to the model. 
Replaces SoftMax activation.

Added the BinaryCE (Binary Cross Entropy) class as a loss function to the layers_activators module and applied it to the model. 
Replaces Categorical Cross Entropy loss. 

Additionally:

Fixed a loss decay calculation error in the OptimizerAdaM class of the optimizers module. 
Replaced the previous loss_decay method with both a pre_update and post_update method that correctly decays learning rate in accordance with training epoch.