ConvLayer object overhaul + test script
Completely overhauled the ConvLayer object both mathematically and structurally. 

Included is the construction of a small neural network utilizing the keras MNIST dataset to test the intersection of the ConvLayer with a DeepLayer, Sigmoid Activation, Binary Cross-Entropy, and a basic optimization scheme. Currently, the model's output tensor isn't formatted properly to learn off of the data, but the individual layers are interacting correctly.